+++ antarest/study/dao/database/database_study_factory_dao.py
@@ -39,8 +39,7 @@ from antarest.study.business.model.thematic_trimming_model import (
 )
 from antarest.study.dao.api.study_factory_dao import StudyFactoryDao
 from antarest.study.dao.database.database_study_dao import DatabaseStudyDao
-from antarest.study.model import STUDY_VERSION_8_3, STUDY_VERSION_9_2
-from antarest.study.storage.utils import StudyMetadataCreation
+from antarest.study.model import STUDY_VERSION_8_3, STUDY_VERSION_9_2, StudyMetadataCreation
 from antarest.study.storage.variantstudy.business.matrix_constants_generator import GeneratorMatrixConstants
