+++ antarest/study/model.py
@@ -777,3 +777,16 @@ class StudyMetadataCopy:
     owner: Identity
     groups: list[Group]
     directory_id: str | None
+
+
+@dataclasses.dataclass(frozen=True)
+class StudyMetadataCreation:
+    id: str
+    version: StudyVersion
+    managed: bool
+    name: str | None = None
+    author: str | None = None
+    editor: str | None = None
+    created_at: datetime | None = None
+    updated_at: datetime | None = None
+    path: Path | None = None
